289-support-forwarded-mail-in-oasis -- refactor to match message store
