+++ packages/server/channels-api/go.mod
@@ -4,6 +4,7 @@ go 1.19
 
 replace github.com/openline-ai/openline-oasis/packages/server/oasis-api => ./../oasis-api
 
+//replace github.com/openline-ai/openline-customer-os/packages/server/message-store-api => ../../../../openline-customer-os/packages/server/message-store-api
 //replace github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module => ../../../../openline-customer-os/packages/server/customer-os-common-module
 //replace github.com/openline-ai/openline-customer-os/packages/server/message-store => ../../../../openline-customer-os/packages/server/message-store
 
@@ -14,6 +15,7 @@ require (
 	github.com/gin-gonic/gin v1.8.2
 	github.com/gorilla/websocket v1.5.0
 	github.com/joho/godotenv v1.4.0
+	github.com/openline-ai/openline-customer-os/packages/server/message-store-api v0.0.0-00010101000000-000000000000
 	github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module v0.0.0-20230127100359-b1bf9fe2fccf
 	github.com/openline-ai/openline-customer-os/packages/server/message-store v0.0.0-20230113175908-2175601d4942
 	github.com/openline-ai/openline-oasis/packages/server/oasis-api v0.0.0-00010101000000-000000000000
