+++ packages/server/channels-api/routes/webchat.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/gin-gonic/gin"
-	ms "github.com/openline-ai/openline-customer-os/packages/server/message-store/proto/generated"
+	ms "github.com/openline-ai/openline-customer-os/packages/server/message-store-api/proto/generated"
 	c "github.com/openline-ai/openline-oasis/packages/server/channels-api/config"
 	"github.com/openline-ai/openline-oasis/packages/server/channels-api/util"
 	o "github.com/openline-ai/openline-oasis/packages/server/oasis-api/proto/generated"
@@ -66,7 +66,7 @@ func AddWebChatRoutes(conf *c.Config, df util.DialFactory, rg *gin.RouterGroup) 
 		log.Printf("Got message from %s", req.Username)
 
 		//Contact the server and print out its response.
-		message := &ms.WebChatInputMessage{
+		message := &ms.InputMessage{
 			Type:       ms.MessageType_WEB_CHAT,
 			Subtype:    ms.MessageSubtype_MESSAGE,
 			Message:    &req.Message,
@@ -78,7 +78,7 @@ func AddWebChatRoutes(conf *c.Config, df util.DialFactory, rg *gin.RouterGroup) 
 		//Store the message in message store
 		msConn := util.GetMessageStoreConnection(c, df)
 		defer util.CloseMessageStoreConnection(msConn)
-		msClient := ms.NewWebChatMessageStoreServiceClient(msConn)
+		msClient := ms.NewMessageStoreServiceClient(msConn)
 
 		ctx := context.Background()
