+++ packages/server/oasis-api/Makefile
@@ -1,6 +1,4 @@
-
 clean:
-	find ent -mindepth 1 -maxdepth 1 -not \( -name '.placeholder' \) -exec rm -r {} \;
 	rm -rf bin
 
 build:
